refactor: new and build contracts
