+++ crates/pop-cli/src/engines/contract_engine.rs
@@ -4,63 +4,14 @@ use duct::cmd;
 use std::path::{Path, PathBuf};
 
 use contract_build::{
-	execute, new_contract_project, BuildArtifacts, BuildMode, ExecuteArgs, Features, ManifestPath,
-	Network, OptimizationPasses, OutputType, Target, UnstableFlags, Verbosity,
-	DEFAULT_MAX_MEMORY_PAGES,
+	execute, BuildArtifacts, BuildMode, ExecuteArgs, Features, ManifestPath, Network,
+	OptimizationPasses, OutputType, Target, UnstableFlags, Verbosity, DEFAULT_MAX_MEMORY_PAGES,
 };
 use contract_extrinsics::{CallExec, DisplayEvents, ErrorVariant, InstantiateExec, TokenMetadata};
 use ink_env::DefaultEnvironment;
 use sp_weights::Weight;
 use subxt::PolkadotConfig as DefaultConfig;
 use subxt_signer::sr25519::Keypair;
-
-/// Create a new smart contract at `target`
-pub fn create_smart_contract(name: String, target: &Path) -> anyhow::Result<()> {
-	// In this code, out_dir will automatically join `name` to `target`,
-	// which is created prior to the call to this function
-	// So we must pass `target.parent()`
-	new_contract_project(&name, target.canonicalize()?.parent())
-}
-
-/// Build a smart contract
-pub fn build_smart_contract(path: &Option<PathBuf>) -> anyhow::Result<()> {
-	// If the user specifies a path (which is not the current directory), it will have to manually
-	// add a Cargo.toml file. If not provided, pop-cli will ask the user for a specific path. or ask
-	// to the user the specific path (Like cargo-contract does)
-	let manifest_path;
-	if path.is_some() {
-		let full_path: PathBuf =
-			PathBuf::from(path.as_ref().unwrap().to_string_lossy().to_string() + "/Cargo.toml");
-		manifest_path = ManifestPath::try_from(Some(full_path))?;
-	} else {
-		manifest_path = ManifestPath::try_from(path.as_ref())?;
-	}
-
-	let args = ExecuteArgs {
-		manifest_path,
-		verbosity: Verbosity::Default,
-		build_mode: BuildMode::Release,
-		features: Features::default(),
-		network: Network::Online,
-		build_artifact: BuildArtifacts::All,
-		unstable_flags: UnstableFlags::default(),
-		optimization_passes: Some(OptimizationPasses::default()),
-		keep_debug_symbols: false,
-		extra_lints: false,
-		output_type: OutputType::Json,
-		skip_wasm_validation: false,
-		target: Target::Wasm,
-		max_memory_pages: DEFAULT_MAX_MEMORY_PAGES,
-		image: Default::default(),
-	};
-
-	// Execute the build and log the output of the build
-	let result = execute(args)?;
-	let formatted_result = result.display();
-	log::success(formatted_result.to_string())?;
-
-	Ok(())
-}
 
 pub fn test_smart_contract(path: &Option<PathBuf>) -> anyhow::Result<()> {
 	cmd("cargo", vec!["test"]).dir(path.clone().unwrap_or("./".into())).run()?;
@@ -177,6 +128,7 @@ pub async fn dry_run_call(
 mod tests {
 	use super::*;
 	use anyhow::{Error, Result};
+	use pop_contracts::create_smart_contract;
 	use std::fs;
 
 	fn setup_test_environment() -> Result<tempfile::TempDir, Error> {
@@ -188,52 +140,6 @@ mod tests {
 		assert!(result.is_ok(), "Contract test environment setup failed");
 
 		Ok(temp_dir)
-	}
-
-	#[test]
-	fn test_contract_create() -> Result<(), Error> {
-		let temp_dir = setup_test_environment()?;
-
-		// Verify that the generated smart contract contains the expected content
-		let generated_file_content =
-			fs::read_to_string(temp_dir.path().join("test_contract/lib.rs"))
-				.expect("Could not read file");
-
-		assert!(generated_file_content.contains("#[ink::contract]"));
-		assert!(generated_file_content.contains("mod test_contract {"));
-
-		// Verify that the generated Cargo.toml file contains the expected content
-		fs::read_to_string(temp_dir.path().join("test_contract/Cargo.toml"))
-			.expect("Could not read file");
-
-		Ok(())
-	}
-
-	#[cfg(feature = "unit_contract")]
-	#[test]
-	fn test_contract_build() -> Result<(), Error> {
-		let temp_contract_dir = setup_test_environment()?;
-
-		let build = build_smart_contract(&Some(temp_contract_dir.path().join("test_contract")));
-		assert!(build.is_ok(), "Result should be Ok");
-
-		// Verify that the folder target has been created
-		assert!(temp_contract_dir.path().join("test_contract/target").exists());
-		// Verify that all the artifacts has been generated
-		assert!(temp_contract_dir
-			.path()
-			.join("test_contract/target/ink/test_contract.contract")
-			.exists());
-		assert!(temp_contract_dir
-			.path()
-			.join("test_contract/target/ink/test_contract.wasm")
-			.exists());
-		assert!(temp_contract_dir
-			.path()
-			.join("test_contract/target/ink/test_contract.json")
-			.exists());
-
-		Ok(())
 	}
 
 	#[cfg(feature = "unit_contract")]
