+++ crates/pop-contracts/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "pop-contracts"
+description = "Library for generating,building, deploying and calling ink! smart contracts."
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+contract-build = { version = "4.0.2" }
+
+[dev-dependencies]
+tempfile = "3.8"
