+++ crates/pop-contracts/src/lib.rs
@@ -1,0 +1,5 @@
+mod build;
+mod new;
+
+pub use build::build_smart_contract;
+pub use new::create_smart_contract;
