+++ crates/pop-parachains/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "pop-parachains"
+description = "Library for generating, building and running parachains."
+version = "0.0.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
